Extract ZoneView to its own component
